feat: add field selector indexing
Any type that implements the Fields interface will have those field
names added as columns to the database. This should significantly
improve the performance of lists with field selectors.

If greater performance is desired, then the types that implement the
FieldIndexer interface will have the appropriate index added for the
table.

Signed-off-by: Donnie Adams <donnie@acorn.io>